Task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 int main(void)
 {
 	int num1;
-	unsigned long int num2, num3, num4;
+	unsigned long long int num2, num3, num4;
 
 	num2 = 0;
 	num3 = 1;
 	for (num1 = 0; num1 < 98; num1++)
 	{
 		num4 = num2 + num3;
-		printf("%lu", num4);
+		printf("%llu", num4);
 		if (num1 < 97)
 		{
 			printf(", ");
